sum_columns_2darray.c
INSTRUCTIONS: In a 2-dimensional array of size 3X3, input the elements in the array. Call the name of the array nineElements. In the second array called threeElements, the first element, the value is the sum of the elements in the first column. The second element, the value is the sum of the elements in the second column. The third element, the value is the sum of the elements in the third column.